refactor(sessions): split session-view store into base slice + concrete store
Extract createBaseSessionViewSlice + SessionViewContext + useSessionViewBaseStore
from the monolithic store.ts (now store/index.tsx) mirroring trace-view's pattern.
Shared children (list, trace-item, span-panel) consume the base hook; the concrete
SessionViewStore keeps session/search/timeline state and dual-provides both contexts.
Pure refactor — regular session view behavior unchanged.
